+++ src/apps/ScopeChangeRequest/Components/Workflow/Atoms/signingAtom.ts
@@ -1,0 +1,11 @@
+import { Atom } from '@dbeining/react-atom';
+import { CriteriaSignState } from '../../../types/scopeChangeRequest';
+
+export interface SigningAction {
+    buttonText: string;
+    action: CriteriaSignState | 'Reassign';
+    criteriaId: string;
+    stepId: string;
+}
+
+export const actionWithCommentAtom = Atom.of<SigningAction | null>(null);
