+++ src/apps/ScopeChangeRequest/Components/Workflow/Contributor/contributor.styles.ts
@@ -1,0 +1,15 @@
+import styled from 'styled-components';
+
+export const CommentFieldWrapper = styled.div`
+    margin: 0.4rem 0;
+`;
+
+export const FlexColumn = styled.div`
+    display: flex;
+    flex-direction: column;
+`;
+
+export const ButtonContainer = styled.span`
+    display: flex;
+    align-items: center;
+`;
