--- src/apps/ScopeChangeRequest/Components/Workflow/Contributor/index.ts
@@ -1,1 +1,0 @@
-export * from './Components/Contributors'
