--- src/apps/ScopeChangeRequest/Components/Workflow/Step/index.ts
@@ -1,1 +1,0 @@
-export * from './Components/WorkflowStep'
