+++ src/apps/ScopeChangeRequest/Components/Workflow/Utils/getCriteriaStatus.ts
@@ -1,0 +1,12 @@
+import { CriteriaStatus } from '../Criteria/Components/CriteriaDetail';
+import { Criteria } from '../../../types/scopeChangeRequest';
+
+export function getCriteriaStatus(criteria: Criteria, isCurrentStep: boolean): CriteriaStatus {
+    if (!criteria.signedState) {
+        return isCurrentStep ? 'Active' : 'Inactive';
+    } else if (criteria.signedState === 'Approved') {
+        return 'Approved';
+    } else {
+        return 'Rejected';
+    }
+}
